requirements.txt
Django==1.11
djangorestframework==3.6.2
httpie==0.9.9
Pygments==2.2.0
pytz==2017.2
requests==2.13.0
